requirements.txt
### FILE: requirements.txt

transformers>=4.30.0
datasets>=2.0.0
torch>=1.13.0
tqdm
seqeval
spacy
python-dateutil
fuzzywuzzy
python-Levenshtein

